backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pandas>=2.1.0
numpy>=1.26.0
python-multipart>=0.0.6
openpyxl>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
openai>=1.10.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pyjwt>=2.8.0
pdfplumber>=0.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
